Fix default collation deserialization
We've run into an issue where the database default collation doesn't
necessarily exist as a "real" collation in pg_collation. This code
specializes the deserialization code to handle this: if we don't find
the collation we're trying to deserialize, check if the name matches
that of the default collation and if so assume it's the default
collation.